Refs #209: Renamed logging settings script
